gsdx-ogl: use integer for blend factor
Integer argument&comparison might be lighter

V2: Forget to change one OMSetBlendState call

+++ plugins/GSdx/GLState.h
@@ -34,7 +34,7 @@ namespace GLState {
 	extern GLenum f_sRGB;
 	extern GLenum f_dRGB;
 	extern uint32 wrgba;
-	extern float bf;
+	extern int bf;
 
 	extern bool depth;
 	extern GLenum depth_func;
